Refactoring in common.app.models
