+++ common/app/model/PressedElements.scala
@@ -1,0 +1,9 @@
+package model.pressed
+
+import model.content.{MediaAtom}
+import model.{VideoElement}
+
+final case class PressedElements(
+  mainVideo: Option[VideoElement],
+  mediaAtoms: Seq[MediaAtom]
+)
