+++ common/app/model/PressedFields.scala
@@ -1,0 +1,7 @@
+package model.pressed
+
+final case class PressedFields(
+  main: String,
+  body: String,
+  standfirst: Option[String]
+)
